Split frontend context hooks
