+++ frontend/app/pages/CreateTournamentPage.tsx
@@ -10,7 +10,7 @@ import {
 } from '@/components/ui/breadcrumb'
 import { Button } from '@/components/ui/button'
 import { Spinner } from '@/components/ui/spinner'
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuth } from '@/contexts/useAuth'
 import loc from '@/lib/locales'
 
 export default function CreateTournamentPage() {
