+++ frontend/app/pages/SessionPage.tsx
@@ -32,9 +32,9 @@ import {
 } from '@/components/ui/select'
 import { Spinner } from '@/components/ui/spinner'
 import UserRow from '@/components/user/UserRow'
-import { useAuth } from '@/contexts/AuthContext'
-import { useConnection } from '@/contexts/ConnectionContext'
-import { useData } from '@/contexts/DataContext'
+import { useAuth } from '@/contexts/useAuth'
+import { useConnection } from '@/contexts/useConnection'
+import { useData } from '@/contexts/useData'
 import loc from '@/lib/locales'
 import type { SessionWithSignups } from '@common/models/session'
 import { isUpcoming } from '@common/utils/date'
