+++ frontend/app/pages/UsersPage.tsx
@@ -22,8 +22,8 @@ import { Item, ItemContent, ItemMedia } from '@/components/ui/item'
 import { Skeleton } from '@/components/ui/skeleton'
 import UserForm from '@/components/user/UserForm'
 import UserRow from '@/components/user/UserRow'
-import { useAuth } from '@/contexts/AuthContext'
-import { useData } from '@/contexts/DataContext'
+import { useAuth } from '@/contexts/useAuth'
+import { useData } from '@/contexts/useData'
 import loc from '@/lib/locales'
 import type { UserInfo } from '@common/models/user'
 import { PlusIcon } from '@heroicons/react/24/solid'
