+++ frontend/components/match/MatchRow.tsx
@@ -1,6 +1,6 @@
-import { useAuth } from '@/contexts/AuthContext'
-import { useConnection } from '@/contexts/ConnectionContext'
-import { useData } from '@/contexts/DataContext'
+import { useAuth } from '@/contexts/useAuth'
+import { useConnection } from '@/contexts/useConnection'
+import { useData } from '@/contexts/useData'
 import loc from '@/lib/locales'
 import type { EditMatchRequest, Match } from '@common/models/match'
 import type { UserInfo } from '@common/models/user'
