+++ frontend/components/session/SessionForm.tsx
@@ -1,5 +1,5 @@
-import { useAuth } from '@/contexts/AuthContext'
-import { useConnection } from '@/contexts/ConnectionContext'
+import { useAuth } from '@/contexts/useAuth'
+import { useConnection } from '@/contexts/useConnection'
 import loc from '@/lib/locales'
 import type { SessionWithSignups } from '@common/models/session'
 import { useState, type ComponentProps, type SubmitEvent } from 'react'
