+++ frontend/components/session/SessionsList.tsx
@@ -1,7 +1,7 @@
 import { Empty } from '@/components/ui/empty'
 import { Skeleton } from '@/components/ui/skeleton'
-import { useAuth } from '@/contexts/AuthContext'
-import { useData } from '@/contexts/DataContext'
+import { useAuth } from '@/contexts/useAuth'
+import { useData } from '@/contexts/useData'
 import loc from '@/lib/locales'
 import type { SessionWithSignups } from '@common/models/session'
 import { Plus } from 'lucide-react'
