+++ frontend/components/timeentries/TimeEntryInput.tsx
@@ -32,9 +32,9 @@ import {
 import { useLocation } from 'react-router'
 import { toast } from 'sonner'
 import { twMerge } from 'tailwind-merge'
-import { useAuth } from '../../contexts/AuthContext'
-import { useConnection } from '../../contexts/ConnectionContext'
-import { useData } from '../../contexts/DataContext'
+import { useAuth } from '../../contexts/useAuth'
+import { useConnection } from '../../contexts/useConnection'
+import { useData } from '../../contexts/useData'
 import { TextField } from '../FormFields'
 import { SessionRow } from '../session/SessionRow'
 import { TrackRow } from '../track/TrackRow'
