+++ frontend/components/tournament/TournamentCard.tsx
@@ -1,6 +1,6 @@
-import { useAuth } from '@/contexts/AuthContext'
-import { useConnection } from '@/contexts/ConnectionContext'
-import { useData } from '@/contexts/DataContext'
+import { useAuth } from '@/contexts/useAuth'
+import { useConnection } from '@/contexts/useConnection'
+import { useData } from '@/contexts/useData'
 import loc from '@/lib/locales'
 import type { TournamentBracket } from '@backend/database/schema'
 import type {
