+++ frontend/components/user/UserForm.tsx
@@ -1,5 +1,5 @@
-import { useAuth } from '@/contexts/AuthContext'
-import { useConnection } from '@/contexts/ConnectionContext'
+import { useAuth } from '@/contexts/useAuth'
+import { useConnection } from '@/contexts/useConnection'
 import loc from '@/lib/locales'
 import { type UserInfo } from '@common/models/user'
 import { useState, type ComponentProps, type SubmitEvent } from 'react'
