+++ frontend/contexts/useAuth.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react'
+import { AuthContext } from './auth-context'
+
+export function useAuth() {
+  const context = useContext(AuthContext)
+  if (!context) throw new Error('useAuth must be used inside AuthProvider')
+  return context
+}
