+++ frontend/contexts/useData.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react'
+import { DataContext } from './data-context'
+
+export const useData = () => {
+  const context = useContext(DataContext)
+  if (!context) throw new Error('useData must be used inside DataProvider')
+  return context
+}
